Debug ListImagesServlet, UploadServlet
