+++ portfolio/src/main/java/com/google/sps/servlets/ListImagesServlet.java
@@ -15,7 +15,12 @@ import javax.servlet.http.HttpServletResponse;
 @WebServlet("/list-images")
 public class ListImagesServlet extends HttpServlet {
 
-  @Override
+    /**
+     *
+     */
+    private static final long serialVersionUID = 1L;
+
+    @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
     // List all of the uploaded files.
     String projectId = "rshillingford-sps-spring21";
